:arrow_right: Move Region enum to /common namespace
